requirements.txt
click==6.6
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.0
Flask==0.11.1
Flask-SocketIO==2.8.1
greenlet==0.4.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-engineio==1.1.0
python-rtmidi==1.0.0
python-socketio==1.6.1
six==1.10.0
Werkzeug==0.11.11
